src/webapp/googlegemini/requirements.txt
Django==5.0.4
django-tables2==2.7.0
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
Pillow==10.3.0
google-generativeai==0.5.2
googlemaps==4.10.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
djangorestframework==3.15.1
django-cors-headers==4.3.1